requirements.txt
python
django
djangorestframework
django-extensions
django-filter
django-cors-headers
pyjwt
